Add failing test for TimestampedValue PCollection
A PCollection of TimestampedValues can be created with a DoFn that
adds the timestamp info along with the value in a TimestampedValue
object, in a ParDo PTransform. But, this is not possible because
when the runner evaluates the ParDo, it converts the element of
TimestampedValue to a WindowedValue with the same value and
timestamp as the TimestampedValue object. But, this value used is
not the TimestampedValue object, which should be the case.